Separate extension to different files
